implements OptionsOverlay component
